Create Helix clients on demand
It turns out that I cannot have a global Helix client, because some
times I will want to use it using an app access token (to fetch
information from a stream), but other times I will want to use it on
behalf on an user (to issue a ban or to create a chat token).

In this commit, I replace the global HelixApi client with a new set of
builders. Forget about the old code, the new helixapi package works like
this:

* There is a builder, which is a class that constructs helix.Client
  objects, fully initialised, with the client ID, the client secret and
  either the app access token or the user access token.

* To provide the access token, there is a TokenProvider, allowing me to
  inject the _strategy_ to use in order to assign a token to the
  helix.Client:
  - Currently there is an AppAccessTokenProvider strategy which simply
    slaps an access token into the helix.Client after using the Twitch
    API to fetch it.
  - The next step is to create an UserAccessTokenProvider strategy which
    is more elaborate, gets the access token or the refresh token from a
    keystore, slaps the access token into the helix.Client, maybe asks
    the user to log in.

* But because the builder client ID and secret are constants provided in
  the static config, there is a builder for builders, which is the
  actual helixapi.HelixApi module. This is the one that gets injected by
  fx, and it is possible to request to this builder a new
  helix.Client backed by an access token or by an user token.
  It orchestrates the construction of the helix.Client using the
  appropiate TokenProvider strategy.